Moves crack methods to CrackModule
