+++ test/crack_module_test.rb
@@ -2,11 +2,7 @@ require "./test/test_helper"
 
 class CrackModuleTest < MiniTest::Test
   def setup
-    @enigma = Enigma.new
-    @enigma.shifts.create_keys(12345)
-    @enigma.shifts.offset_integrated("032489")
-    @enigma.shifts.create_shifts
-    @enigma.shifts.create_shifted_arrays(12345,"032489")
+    @cracker = Cracker.new
   end
 
   def test_cracker_finds_key_for_encrypted_message
